+++ Sabv/Services/Sabv.Services.Data/MessagesService.cs
@@ -1,5 +1,6 @@
 ﻿namespace Sabv.Services.Data
 {
+    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
@@ -19,6 +20,16 @@
 
         public async Task AddAsync(string content, ApplicationUser user)
         {
+            if (string.IsNullOrEmpty(content))
+            {
+                throw new ArgumentNullException("Message content cannot be null or empty.");
+            }
+
+            if (user == null)
+            {
+                throw new ArgumentNullException("User cannot be null.");
+            }
+
             await this.messagesRepo.AddAsync(new Message()
             {
                 Content = content,
@@ -32,6 +43,12 @@
         public async Task DeleteAsync(int id)
         {
             var entity = this.messagesRepo.All().FirstOrDefault(x => x.Id == id);
+
+            if (entity == null)
+            {
+                throw new ArgumentNullException("Entity with given id not found.");
+            }
+
             this.messagesRepo.Delete(entity);
             await this.messagesRepo.SaveChangesAsync();
         }
